everything will break in 3...2.. (2)

+++ Simple REST client/src/main/java/vadim/andreich/api/model/Telegram.java
@@ -17,7 +17,7 @@ public class Telegram {
     @Id
     @Column(name = "chat_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long chatId;
+    private int chatId;
 
     @Column(name = "username")
     private String username;
